+++ mitumbaesales/api/serializers.py
@@ -1,0 +1,18 @@
+from rest_framework import serializers
+from orders.models import Order, OrderItem
+
+class OrderSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Order
+        fields = '__all__'
+
+class OrderItemSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = OrderItem
+        fields = '__all__'
+
+
+
+
+
+
