+++ mitumbaesales/api/views.py
@@ -1,3 +1,14 @@
-from django.shortcuts import render
+from rest_framework import generics
+from orders.models import Order, OrderItem
+from .serializers import OrderSerializer, OrderItemSerializer
 
-# Create your views here.
+class OrderListCreateAPIView(generics.ListCreateAPIView):
+    queryset = Order.objects.all()
+    serializer_class = OrderSerializer
+
+class OrderItemListCreateAPIView(generics.ListCreateAPIView):
+    queryset = OrderItem.objects.all()
+    serializer_class = OrderItemSerializer
+
+
+
