+++ mitumbaesales/mitumbaesales/urls.py
@@ -19,5 +19,7 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('reviews/', include('reviews.urls')),
+    path('api/', include('api.urls')), 
 ]
+
+
