+++ mitumbaesales/orders/admin.py
@@ -1,3 +1,11 @@
 from django.contrib import admin
 
-# Register your models here.
+
+
+from .models import Order
+admin.site.register(Order)
+
+from .models import OrderItem
+admin.site.register(OrderItem)
+# admin.site.register(Product)
+
