+++ mitumbaesales/payments/models.py
@@ -4,7 +4,7 @@ from django.db import models
 # Create your models here.
 import uuid
 from authentication.models import User
-from order.models import Order
+from orders.models import Order
 class Payment(models.Model):
     PAYMENT_METHOD_CHOICES = [
         ('M-Pesa', 'M-Pesa'),
