+++ mitumbaesales/product/models.py
@@ -15,7 +15,7 @@ class Product(models.Model):
     name = models.CharField(max_length=100)
     price = models.DecimalField(max_digits=10, decimal_places=2)
     stock_quantity = models.PositiveIntegerField(default=1)
-    image = models.ImageField(upload_to='product_images/')
+    image = models.URLField(max_length=400)
     description = models.TextField()
     size = models.CharField(max_length=10)
     created_at = models.DateTimeField(auto_now_add=True)
